+++ src/api/routes/donateTab/index.js
@@ -1,6 +1,7 @@
 import express from "express";
 import pool from "../../../db/index.js";
 import auth from "../../../middleware/auth.js";
+import deleteFile from "../../helpers/deleteMedia.js";
 const router = express.Router();
 
 router.post("/", auth, async (req, res) => {
@@ -45,6 +46,20 @@ router.put("/", auth, async (req, res) => {
     connection = await pool.getConnection();
     await connection.beginTransaction();
 
+    const [header] = await connection.query(
+      "SELECT * FROM donateTabs WHERE id = ?",
+      [id]
+    );
+
+    if (header.length === 0) {
+      return res.status(200).json({ msg: "Logo not found" });
+    }
+
+    const filePath = header[0].logo;
+    if (filePath && filePath !== logo) {
+      await deleteFile(filePath);
+    }
+
     const [result] = await connection.query(
       `UPDATE donateTabs SET logo = ?, updatedAt = CURRENT_TIMESTAMP WHERE id = ?`,
       [logo, id]
@@ -74,6 +89,20 @@ router.delete("/", auth, async (req, res) => {
     const { id } = req.query;
     connection = await pool.getConnection();
     await connection.beginTransaction();
+
+    const [header] = await connection.query(
+      "SELECT * FROM donateTabs WHERE id = ?",
+      [id]
+    );
+
+    if (header.length === 0) {
+      return res.status(200).json({ msg: "Logo not found" });
+    }
+
+    const filePath = header[0].logo;
+    if (filePath) {
+      await deleteFile(filePath);
+    }
 
     const [result] = await connection.query(
       `DELETE FROM donateTabs WHERE id = ?`,
